backend/containers/auth-service/requirements.txt
pyjwt
aiosqlite
fastapi
uvicorn
pydantic[email]